+++ src/util.h
@@ -177,6 +177,8 @@ void FileCommit(FILE *fileout);
 
 int GetFilesize(FILE* file);
 
+std::string TimestampToHRDate(double dtm);
+
 bool RenameOver(boost::filesystem::path src, boost::filesystem::path dest);
 boost::filesystem::path GetDefaultDataDir();
 boost::filesystem::path GetProgramDir();
